scripts/python/requirements.txt
sense-emu==1.2.1
mysql-connector==2.2.9
matplotlib==3.9.2